Change app to accept list items intead items

+++ lib/src/domain/services/verse_service.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'package:daily_messages/src/domain/http/http_client.dart';
 import 'package:daily_messages/src/domain/models/bible_verse_model.dart';
 import 'package:daily_messages/src/domain/repositories/bible_verse_repository.dart';
+import 'package:daily_messages/src/domain/services/check_internet_connection_service.dart';
 import 'package:daily_messages/src/domain/services/shared_preferences_service.dart';
 import 'package:daily_messages/src/domain/stores/bible_verse_store.dart';
 import 'package:daily_messages/src/ui/widgets/custom_snack_bar.dart';
@@ -19,45 +20,62 @@ class VerseService {
   bool _isFetchingMore = false;
   final SharedPreferencesService _sharedPreferencesService =
       SharedPreferencesService();
+  final CheckInternetConnectionService _checkInternetConnectionService =
+      CheckInternetConnectionService();
   final today = DateFormat('yyyy-MM-dd').format(DateTime.now());
 
   Future<List<BibleVerseModel>> getVersesData(BuildContext context) async {
-    final lastDataSaved =
-        await _sharedPreferencesService.getSharedData('lastDataSaved') ?? '';
-    await Future.delayed(const Duration(milliseconds: 100));
+    try {
+      final isConnected =
+          await _checkInternetConnectionService.checkInternetConnection();
 
-    if (lastDataSaved != today) {
-      await _sharedPreferencesService.removeSharedData('bibleVerses');
-      await _sharedPreferencesService.removeSharedData('lastDataSaved');
-    }
-    final recoveredData = await recoveryData();
-    await Future.delayed(const Duration(milliseconds: 100));
+      final lastDataSaved =
+          await _sharedPreferencesService.getSharedData('lastDataSaved') ?? '';
+      await Future.delayed(const Duration(milliseconds: 100));
 
-    if (recoveredData.isNotEmpty) {
-      return recoveredData;
-    } else {
-      if (!context.mounted) return [];
-      return await fetchData(context);
+      if (lastDataSaved != today && isConnected) {
+        await _sharedPreferencesService.removeSharedData('bibleVerses');
+        await _sharedPreferencesService.removeSharedData('lastDataSaved');
+      }
+      final recoveredData = await recoveryData();
+      await Future.delayed(const Duration(milliseconds: 100));
+
+      if (recoveredData.isNotEmpty) {
+        return recoveredData;
+      } else {
+        if (!context.mounted) return [];
+        return await fetchData(context);
+      }
+    } catch (e) {
+      return [];
     }
   }
 
   Future<List<BibleVerseModel>> getMoreVersesData(BuildContext context) async {
-    final lastDataSaved =
-        await _sharedPreferencesService.getSharedData('lastDataSaved') ?? '';
-    await Future.delayed(const Duration(milliseconds: 100));
+    try {
+      final isConnected =
+          await _checkInternetConnectionService.checkInternetConnection();
 
-    if (lastDataSaved != today) {
-      await _sharedPreferencesService.removeSharedData('bibleVerses');
-      await _sharedPreferencesService.removeSharedData('lastDataSaved');
-    }
-    final recoveredData = await recoveryData();
-    await Future.delayed(const Duration(milliseconds: 100));
+      final lastDataSaved =
+          await _sharedPreferencesService.getSharedData('lastDataSaved') ?? '';
+      await Future.delayed(const Duration(milliseconds: 100));
 
-    if (recoveredData.isNotEmpty && recoveredData.length == 100) {
-      return recoveredData;
-    } else {
-      if (!context.mounted) return [];
-      return await fetchData(context);
+      if (lastDataSaved != today && isConnected) {
+        await _sharedPreferencesService.removeSharedData('bibleVerses');
+        await _sharedPreferencesService.removeSharedData('lastDataSaved');
+      }
+      final recoveredData = await recoveryData();
+      await Future.delayed(const Duration(milliseconds: 100));
+
+      if (recoveredData.isNotEmpty && recoveredData.length >= 100) {
+        await saveVerse(_bibleVerseModels);
+        return recoveredData;
+      } else {
+        if (!context.mounted) return [];
+        return await fetchData(context);
+      }
+    } catch (e) {
+      return [];
     }
   }
 
@@ -65,12 +83,12 @@ class VerseService {
     try {
       if (_bibleVerseModels.isEmpty || !_isFetchingMore) {
         _isFetchingMore = true;
-        await _store.getMultipleVerses(10, context);
+        await _store.getMultipleVerses(1, context);
         _bibleVerseModels.addAll(_store.state.value);
         _isFetchingMore = false;
       }
       await saveVerse(_bibleVerseModels);
-      return _bibleVerseModels;
+      return _store.state.value;
     } catch (e) {
       _isFetchingMore = false;
       if (!context.mounted) return [];
